In Memory Identity for ASPNET4 MVC5.
This is for the times that an external user database is already in place and you only need to get authenticated.
all you need is that unique ID back from the external IDP.

In this case it is just twitter for the test.

I want to use the OWIN way of doing things, which is really just letting Microsoft do what it want via its example apps.
This is just fooling the framework into thinking that we actually did database work when an new user came in.

Look for CustomUser, CustomSignInManager and CustomUserManager

BTW: I left everything intact from the wizard app and only did the following
1. Replace ApplicationUser,ApplicationSignInManager, and ApplicationUserManager with the Custom Varients.

Startup.ConfigureAuth
 /*
            app.CreatePerOwinContext(ApplicationDbContext.Create);
            app.CreatePerOwinContext<ApplicationUserManager>(ApplicationUserManager.Create);
            app.CreatePerOwinContext<ApplicationSignInManager>(ApplicationSignInManager.Create);
            */
            app.CreatePerOwinContext(ApplicationDbContext.Create);
            app.CreatePerOwinContext<CustomUserManager>(CustomUserManager.Create);
            app.CreatePerOwinContext<CustomSignInManager>(CustomSignInManager.Create);

2. Make a fake email if one doesn't exist from the claims and skip the whole register email thing.  this is done in the AccountController.ExternalLoginCallback function.