Got everything to work.
So the secret is to just use ClientCredential and implement my own CustomClaimsProvider.
In the CustomClaimsProvider I would be evaluating the extra params and generate my own claims based upon that.

NOTE: ClientCredentials vs CustomGrant.

ClientCredentials: Simple and I don't need a USER store.
CustomGrant: in this scenario a user store is required and it is assumed that the custom grant is against said user.

Seems to me that in both cases you have to pass custom params and evaluation of those params still fires code that basically results in claims being set.  I can do the same thing with simple ClientCredentials and don't need a User store.